Adjusted integration tests Azure Active Directory authentication
Instead of manually emitting and parsing the ROPC grant flow, we now leverage the `UsernamePasswordCredential` from the `Azure.Identity` package. Since this credential makes use of a public client application, the application registration in AAD has been adjusted and the client secret is no longer required. The `UserAuthenticationSettings` class, the custom `TokenResponse` class as well as the GitHub action running the tests have been adjusted or removed accordingly.

In order to reduce the repetitive setup of the authenticated and anonymous `HttpClient`s required in the endpoint tests, a new test base class `EndpointTestBase` containing both clients has been introduced.