feat(time): Replace custom `IDateTimeProvider` with .NET 8 built-in `TimeProvider`
Removed the custom `IDateTimeProvider` interface and its implementation, `DateTimeProvider`, from the code and replaced it with the new .NET 8 built-in `TimeProvider`.

The change was made to simplify the architecture and reduce code that was mostly used for testing purposes. Now, with the `TimeProvider`, we are still able to retrieve the current time using `GetUtcNow()`. These changes also reflect on removing the `IDateTimeProvider` registration from `Program.cs`.