refactor(mapping): extracted mapping logic to extension methods (#492)
Moved the mapping logic from the individual endpoints to extension
methods. This was done to reduce the complexity, reduce code duplication
and improve the maintainability of the codebase.

Also, removed the duplicate `userprefs` from `.gitignore`.

Furthermore, the `ValidationErrors.ResponseBuilder` method was extracted
to another class to enhance code readability and segregation.