chore(terraform): remove hard-coded Terraform Cloud config (#563)
The configuration for Terraform Cloud has been refactored from being
hard-coded in the infra provider files to being received as user inputs
in the GitHub Actions workflows. The Terraform Cloud organization and
workspace are now required inputs for the terraform-apply and
terraform-plan actions. Additionally, the Terraform required version has
been updated from `>= 1.4.0` to `>= 1.6.0`.