Update telemetry and optimize dependencies
The telemetry settings were updated across the infrastructure and the NoPlan.Api project. This change includes the addition of the `OTEL_SERVICE_NAME` in the environment variables and container app configuration. Moreover, the list of project dependencies was cleaned up, which involved package upgrades and the removal of unused packages, particularly those related to OpenTelemetry.